Move runtime configuration class to runtime artifact
